pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "earthengine-api>=1.6.6",
    "fastapi>=0.116.1",
    "folium>=0.20.0",
    "google-auth>=2.40.3",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
]
